Creating 1st and 2nd API methods
