+++ employees/index.html
@@ -4,6 +4,7 @@
       <title>Generator</title>
       <link href='style.css' rel='stylesheet' type='text/css'>
       <script src='generator.js'></script>
+      <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
    </head>
    <body>
       <h1>Employee skills report generator</h1>
